refactor(auth): centralize login/session helpers in utils, keep flashes when clearing session, and add logout confirmation flash
